move hackathon to correct place
